Move entrypoint to TypeScript
